+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v3.0.18
+
+- Add more parameters to API type configuration
+- Extract model for HA DeviceInfo from API
+- Add translation key to all components with name ready for translation
+
 ## v3.0.17
 
 - Fix logic to extract protocol codes